build library
