+++ .gitignore
@@ -10,6 +10,7 @@ graph_subst_single.pb
 /tmp
 bert_time.txt
 bert_old_time.txt
+tensatcpp
 *.pdf
 *.png
 ._*
