+++ Cargo.toml
@@ -4,6 +4,9 @@ version = "0.1.0"
 authors = ["Remy Wang <remywang@uw.edu>"]
 edition = "2018"
 
+[lib]
+crate-type = ["staticlib", "lib"]
+
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
