+++ src/graph.cc
@@ -1,5 +1,5 @@
-#include "tensat/src/input.rs.h"
 #include "rust/cxx.h"
+#include "tensat/src/input.rs.h"
 #include <utility>
 #include <bits/stdc++.h>
